- skipped camera

+++ drivers/char/diag/diag_pcie.c
@@ -61,6 +61,9 @@ static void diag_pcie_event_notifier(struct mhi_dev_client_cb_reason *reason)
 		pcie_info = &diag_pcie[i];
 		if (reason->reason == MHI_DEV_TRE_AVAILABLE)
 			if (reason->ch_id == pcie_info->in_chan) {
+				DIAG_LOG(DIAG_DEBUG_MUX,
+					"diag: queueing a read off pcie (pcie_info->read_cnt: %d)\n",
+					pcie_info->read_cnt);
 				queue_work(pcie_info->wq,
 					&pcie_info->read_work);
 				break;
@@ -86,17 +89,29 @@ void diag_pcie_read_work_fn(struct work_struct *work)
 	ureq.buf = pcie_info->in_chan_attr.read_buffer;
 	ureq.len = pcie_info->in_chan_attr.read_buffer_size;
 	ureq.transfer_len = 0;
-	bytes_avail = mhi_dev_read_channel(&ureq);
-	if (bytes_avail < 0)
-		return;
-	DIAG_LOG(DIAG_DEBUG_MUX, "read total bytes %d from chan:%d",
-		bytes_avail, pcie_info->in_chan);
-	pcie_info->read_cnt++;
 
-	if (pcie_info->ops && pcie_info->ops->read_done)
-		pcie_info->ops->read_done(pcie_info->in_chan_attr.read_buffer,
-					ureq.transfer_len, pcie_info->ctxt);
+	do {
+		bytes_avail = mhi_dev_read_channel(&ureq);
+		if (bytes_avail < 0) {
+			DIAG_LOG(DIAG_DEBUG_MUX,
+				"diag: failed to read with error: %d\n",
+				bytes_avail);
+			return;
+		}
+		if (bytes_avail == 0)
+			return;
 
+		pcie_info->read_cnt++;
+		DIAG_LOG(DIAG_DEBUG_MUX,
+			"diag: read total bytes: %d from chan: %d with read_cnt: %d\n",
+			bytes_avail, pcie_info->in_chan, pcie_info->read_cnt);
+
+		if (pcie_info->ops && pcie_info->ops->read_done)
+			pcie_info->ops->read_done(
+				pcie_info->in_chan_attr.read_buffer,
+				ureq.transfer_len, pcie_info->ctxt);
+
+	} while (bytes_avail > 0);
 }
 
 static void diag_pcie_buf_tbl_remove(struct diag_pcie_info *pcie_info,
@@ -429,11 +444,12 @@ void diag_pcie_client_cb(struct mhi_dev_client_cb_data *cb_data)
 	case  MHI_STATE_CONNECTED:
 		if (cb_data->channel == pcie_info->out_chan) {
 			DIAG_LOG(DIAG_DEBUG_MUX,
-			"diag: Received connect event from MHI for %d\n",
+				"diag: Received connect event from MHI for %d",
 				pcie_info->out_chan);
 			if (atomic_read(&pcie_info->enabled)) {
 				DIAG_LOG(DIAG_DEBUG_MUX,
-				"diag: pcie channel is already enabled\n");
+				"diag: Channel %d is already enabled",
+				pcie_info->out_chan);
 				return;
 			}
 			queue_work(pcie_info->wq, &pcie_info->open_work);
@@ -442,11 +458,12 @@ void diag_pcie_client_cb(struct mhi_dev_client_cb_data *cb_data)
 	case MHI_STATE_DISCONNECTED:
 		if (cb_data->channel == pcie_info->out_chan) {
 			DIAG_LOG(DIAG_DEBUG_MUX,
-			"diag: Received disconnect event from MHI for %d",
+				"diag: Received disconnect event from MHI for %d",
 				pcie_info->out_chan);
 			if (!atomic_read(&pcie_info->enabled)) {
 				DIAG_LOG(DIAG_DEBUG_MUX,
-				"diag: pcie channel is already disabled\n");
+				"diag: Channel %d is already disabled",
+				pcie_info->out_chan);
 				return;
 			}
 			queue_work(pcie_info->wq, &pcie_info->close_work);
@@ -498,12 +515,9 @@ static void diag_pcie_connect(struct diag_pcie_info *ch)
 	queue_work(ch->wq, &(ch->read_work));
 }
 
-void diag_pcie_open_work_fn(struct work_struct *work)
+static void diag_pcie_open_channels(struct diag_pcie_info *pcie_info)
 {
 	int rc = 0;
-	struct diag_pcie_info *pcie_info = container_of(work,
-						      struct diag_pcie_info,
-						      open_work);
 
 	if (!pcie_info || atomic_read(&pcie_info->enabled))
 		return;
@@ -546,6 +560,15 @@ handle_in_err:
 handle_not_rdy_err:
 	mutex_unlock(&pcie_info->in_chan_lock);
 	mutex_unlock(&pcie_info->out_chan_lock);
+}
+
+void diag_pcie_open_work_fn(struct work_struct *work)
+{
+	struct diag_pcie_info *pcie_info = container_of(work,
+						      struct diag_pcie_info,
+						      open_work);
+
+	diag_pcie_open_channels(pcie_info);
 }
 
 /*
@@ -712,7 +735,10 @@ int diag_pcie_register(int id, int ctxt, struct diag_mux_ops *ops)
 	mutex_init(&ch->in_chan_lock);
 	mutex_init(&ch->out_chan_lock);
 	rc = diag_register_pcie_channels(ch);
-	if (rc < 0) {
+	if (rc == -EEXIST) {
+		diag_pcie_open_channels(ch);
+		pr_err("diag: Handled -EEXIST error\n");
+	} else if (rc < 0 && rc != -EEXIST) {
 		if (ch->wq)
 			destroy_workqueue(ch->wq);
 		kfree(ch->in_chan_attr.read_buffer);
